Remove the SNS Topic's display name
There seems to be a bug in AWS's SES Notification validation that
rejects SNS Topics that have a Display Name with spaces, dots, etc.
(even though SNS allows those in the display name). Removing the Display
Name here, since it does not benefit us if it has to match the same
naming constraints as the SNS Topic Name.